+++ lib/src/retry.dart
@@ -4,13 +4,14 @@ import 'package:logging/logging.dart';
 
 import '../youtube_explode_dart.dart';
 
+final _logger = Logger('YoutubeExplode.Retry');
+
 /// Run the [function] each time an exception is thrown until the retryCount
 /// is 0.
 Future<T> retry<T>(
   YoutubeHttpClient? client,
   FutureOr<T> Function() function,
 ) async {
-  final logger = Logger('YoutubeExplode.Retry');
   var retryCount = 5;
 
   // ignore: literal_only_boolean_expressions
@@ -22,7 +23,7 @@ Future<T> retry<T>(
       if (client != null && client.closed) {
         throw HttpClientClosedException();
       }
-      logger.warning('Retrying after exception: $e', e, s);
+      _logger.warning('Retrying after exception: $e', e, s);
       retryCount -= getExceptionCost(e);
       if (retryCount <= 0) {
         rethrow;