+++ lib/src/videos/streams/mixins/hls_stream_info.dart
@@ -1,0 +1,6 @@
+
+import '../../../../youtube_explode_dart.dart';
+
+mixin HlsStreamInfo on StreamInfo {
+  int? get audioItag => null;
+}