+++ lib/src/videos/streams/models/video_quality.dart
@@ -37,3 +37,22 @@ enum VideoQuality {
   /// High quality (4320p).
   high4320
 }
+
+extension QString on VideoQuality {
+  String get qualityString {
+    return switch (this) {
+      VideoQuality.unknown => 'Unknown',
+      VideoQuality.low144 => '144p',
+      VideoQuality.low240 => '240p',
+      VideoQuality.medium360 => '360p',
+      VideoQuality.medium480 => '480p',
+      VideoQuality.high720 => '720p',
+      VideoQuality.high1080 => '1080p',
+      VideoQuality.high1440 => '1440p',
+      VideoQuality.high2160 => '2160p',
+      VideoQuality.high2880 => '2880p',
+      VideoQuality.high3072 => '3072p',
+      VideoQuality.high4320 => '4320p',
+    };
+  }
+}