+++ lib/src/videos/streams/stream_controller.dart
@@ -1,4 +1,5 @@
 import '../../reverse_engineering/dash_manifest.dart';
+import '../../reverse_engineering/hls_manifest.dart';
 import '../video_controller.dart';
 
 class StreamController extends VideoController {
@@ -6,5 +7,9 @@ class StreamController extends VideoController {
 
   Future<DashManifest> getDashManifest(String url) async {
     return DashManifest.parse(await httpClient.getString(url));
+  }
+
+  Future<HlsManifest> getHlsManifest(String url) async {
+    return HlsManifest.parse(await httpClient.getString(url));
   }
 }