+++ lib/src/videos/youtube_api_client.dart
@@ -50,9 +50,9 @@ class YoutubeApiClient {
             'utcOffsetMinutes': 0
           }
         },
-        "thirdParty": {"embedUrl": "https://www.youtube.com/"},
-        "contentCheckOk": true,
-        "racyCheckOk": true
+        'thirdParty': {'embedUrl': 'https://www.youtube.com/'},
+        'contentCheckOk': true,
+        'racyCheckOk': true
       },
       'https://www.youtube.com/youtubei/v1/player?prettyPrint=false',
       headers: {
@@ -94,6 +94,23 @@ class YoutubeApiClient {
     },
   }, 'https://music.youtube.com/youtubei/v1/player?key=AIzaSyAOghZGza2MQSZkY_zfZ370N-PUdXEo8AI&prettyPrint=false');
 
+  static const androidVr = YoutubeApiClient({
+    'context': {
+      'client': {
+        'clientName': 'ANDROID_VR',
+        'clientVersion': '1.56.21',
+        'deviceModel': 'Quest 3',
+        'osVersion': '12',
+        'osName': 'Android',
+        'androidSdkVersion': '32',
+        'hl': 'en',
+        'timeZone': 'UTC',
+        'utcOffsetMinutes': 0,
+      },
+    },
+  }, 'https://www.youtube.com/youtubei/v1/player?prettyPrint=false');
+
+  @Deprecated('Currently this has been disabled on youtube')
   static const webCreator = YoutubeApiClient({
     'context': {
       'client': {
@@ -117,4 +134,42 @@ class YoutubeApiClient {
       },
     },
   }, 'https://www.youtube.com/youtubei/v1/player?prettyPrint=false');
+
+  /// This client also provide high quality muxed stream in the HLS manifest.
+  /// The streams are in m3u8 format.
+  static const safari = YoutubeApiClient({
+    'context': {
+      'client': {
+        'clientName': 'WEB',
+        'clientVersion': '2.20240726.00.00',
+        'userAgent':
+            'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/15.5 Safari/605.1.15,gzip(gfe)',
+        'hl': 'en',
+        'timeZone': 'UTC',
+        'utcOffsetMinutes': 0,
+      },
+    },
+  }, 'https://www.youtube.com/youtubei/v1/player?prettyPrint=false');
+
+  static const tv = YoutubeApiClient(
+      {
+        'context': {
+          'client': {
+            'clientName': 'TVHTML5',
+            'clientVersion': '7.20240724.13.00',
+            'hl': 'en',
+            'timeZone': 'UTC',
+            'gl': 'US',
+            'utcOffsetMinutes': 0
+          }
+        },
+        "contentCheckOk": true,
+        "racyCheckOk": true
+      },
+      'https://www.youtube.com/youtubei/v1/player?prettyPrint=false',
+      headers: {
+        'Sec-Fetch-Mode': 'navigate',
+        'Content-Type': 'application/json',
+        'Origin': 'https://www.youtube.com',
+      });
 }